requirements.txt
fastapi
uvicorn
gunicorn
sodapy
pandas
google-cloud
google-cloud-bigquery